backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "aiosqlite>=0.22.1",
    "fastapi[all]>=0.128.0",
    "passlib[argon2]>=1.7.4",
    "python-dotenv>=1.2.1",
    "python-jose>=3.5.0",
    "sqlmodel>=0.0.31",
]
